fix: Xcodeプロジェクトにファイルが含まれない問題を修正
- RecentFilesManagerをArchiveExtractor.swiftに統合
- RecentFilesViewをContentView.swiftに統合
- 独立したファイルを削除してビルドエラーを解決
- Xcodeプロジェクトへの手動追加が不要な形に変更
